fix: update property types in DownloadProgress constructor
Adjusted the types of total, delta, and bytesPerSecond from float to int to ensure type consistency and accuracy. This change prevents potential precision issues and aligns with expected data formats.

+++ src/Events/AutoUpdater/DownloadProgress.php
@@ -12,7 +12,7 @@ class DownloadProgress implements ShouldBroadcastNow
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
-    public function __construct(public float $total, public float $delta, public float $transferred, public float $percent, public float $bytesPerSecond) {}
+    public function __construct(public int $total, public int $delta, public int $transferred, public float $percent, public int $bytesPerSecond) {}
 
     public function broadcastOn()
     {
